Separate Firebase client tests by test suites.
